+++ instrumentation-genai/opentelemetry-instrumentation-google-genai/src/opentelemetry/instrumentation/google_genai/instrumentor.py
@@ -18,14 +18,13 @@ from opentelemetry._logs import get_logger_provider
 from opentelemetry.instrumentation.instrumentor import BaseInstrumentor
 from opentelemetry.metrics import get_meter_provider
 from opentelemetry.trace import get_tracer_provider
-from opentelemetry.util.genai.completion_hook import load_completion_hook
+from opentelemetry.util.genai.handler import get_telemetry_handler
 
 from .allowlist_util import AllowList
 from .generate_content import (
     instrument_generate_content,
     uninstrument_generate_content,
 )
-from .otel_wrapper import OTelWrapper
 
 
 class GoogleGenAiSdkInstrumentor(BaseInstrumentor):
@@ -44,7 +43,7 @@ class GoogleGenAiSdkInstrumentor(BaseInstrumentor):
     # Inherited, abstract function from 'BaseInstrumentor'. Even though 'self' is
     # not used in the definition, a method is required per the API contract.
     def instrumentation_dependencies(self) -> Collection[str]:  # pylint: disable=no-self-use
-        return ["google-genai>=1.0.0,<2"]
+        return ["google-genai>=1.32.0,<2"]
 
     def _instrument(self, **kwargs: Any):
         tracer_provider = (
@@ -54,17 +53,13 @@ class GoogleGenAiSdkInstrumentor(BaseInstrumentor):
             kwargs.get("logger_provider") or get_logger_provider()
         )
         meter_provider = kwargs.get("meter_provider") or get_meter_provider()
-        otel_wrapper = OTelWrapper.from_providers(
+        handler = get_telemetry_handler(
             tracer_provider=tracer_provider,
             logger_provider=logger_provider,
             meter_provider=meter_provider,
         )
-        completion_hook = (
-            kwargs.get("completion_hook") or load_completion_hook()
-        )
         self._generate_content_snapshot = instrument_generate_content(
-            otel_wrapper,
-            completion_hook,
+            handler,
             generate_content_config_key_allowlist=self._generate_content_config_key_allowlist,
         )